BUG: No way to control date range for TableFilters component Fixes #13885 - TableFilters uses FilterManager date range if not fallback - mock -> fixtures - fallback variable names made clearer - TestTable folded into table test file
