+++ frontend-react/src/components/DataDashboard/DataDashboardTable/DataDashboardTable.test.tsx
@@ -12,7 +12,7 @@ import {
     mockAppInsightsContextReturnValue,
 } from "../../../contexts/__mocks__/AppInsightsContext";
 import { mockSessionContentReturnValue } from "../../../contexts/__mocks__/SessionContext";
-import { mockFilterManager } from "../../../hooks/filters/mocks/MockFilterManager";
+import { filterManagerFixture } from "../../../hooks/filters/filters.fixtures";
 import { mockUseReceiverDeliveries } from "../../../hooks/network/DataDashboard/__mocks__/UseReceiverDeliveries";
 import { mockUseOrganizationReceivers } from "../../../hooks/network/Organizations/__mocks__/ReceiversHooks";
 import { renderApp } from "../../../utils/CustomRenderUtils";
@@ -21,6 +21,7 @@ import { selectDatesFromRange } from "../../../utils/TestUtils";
 
 const mockReceiverServices = receiverServicesGenerator(5);
 const mockActiveReceiver = mockReceiverServices[0];
+const mockFilterManager = { ...filterManagerFixture };
 
 vi.mock("../../../TelemetryService", async (importActual) => ({
     ...(await importActual<typeof import("../../../TelemetryService")>()),
