--- frontend-react/src/hooks/filters/mocks/MockCursorManager.ts
@@ -1,13 +1,0 @@
-import { CursorManager } from "../UseCursorManager";
-
-export const mockCursorManager: CursorManager = {
-    cursors: {
-        current: "",
-        next: "",
-        history: [],
-    },
-    hasPrev: false,
-    hasNext: false,
-    // eslint-disable-next-line no-console
-    update: () => void 0,
-};
